ai-backend/requirements.txt
Flask==3.1.0
torch==2.5.1
pandas==2.2.3
opencv-python==4.10.0.84
Pillow==11.0.0
requests
